requirements.txt
streamlit
numpy
pandas
scikit-learn==1.6.1
joblib
sentence-transformers
langchain-huggingface
python-dotenv
pdfplumber
